feat: fix IAS max_duration limits and move to modernExtend
