+++ src/devices/adurosmart.ts
@@ -149,20 +149,16 @@ export const definitions: DefinitionWithExtend[] = [
         model: "81868",
         vendor: "AduroSmart",
         description: "Siren",
-        fromZigbee: [fz.battery, fz.ias_wd, fz.ias_enroll, fz.ias_siren],
-        toZigbee: [tz.warning_simple, tz.ias_max_duration, tz.warning],
+        fromZigbee: [fz.battery, fz.ias_enroll, fz.ias_siren],
+        toZigbee: [tz.warning_simple, tz.warning],
+        extend: [m.iasWarningMaxDuration()],
         configure: async (device, coordinatorEndpoint) => {
             const endpoint = device.getEndpoint(1);
             await reporting.bind(endpoint, coordinatorEndpoint, ["genBasic", "ssIasZone", "ssIasWd"]);
             await endpoint.read("ssIasZone", ["zoneState", "iasCieAddr", "zoneId"]);
             await endpoint.read("ssIasWd", ["maxDuration"]);
         },
-        exposes: [
-            e.tamper(),
-            e.warning(),
-            e.numeric("max_duration", ea.ALL).withUnit("s").withValueMin(0).withValueMax(600).withDescription("Duration of Siren"),
-            e.binary("alarm", ea.SET, "ON", "OFF").withDescription("Manual start of siren"),
-        ],
+        exposes: [e.tamper(), e.warning(), e.binary("alarm", ea.SET, "ON", "OFF").withDescription("Manual start of siren")],
     },
     {
         zigbeeModel: ["AD-CTW123001"],
