+++ src/lib/modernExtend.ts
@@ -1937,6 +1937,47 @@ export function iasWarning(args: IasWarningArgs = {}): ModernExtend {
     return {toZigbee, exposes, isModernExtend: true};
 }
 
+export function iasWarningMaxDuration(): ModernExtend {
+    const exposes: Expose[] = [
+        e
+            .numeric("max_duration", ea.ALL)
+            .withUnit("s")
+            .withValueMin(0)
+            .withValueMax(65534)
+            .withValueStep(1)
+            .withDescription("Maximum time that the alarm will be active"),
+    ];
+
+    // previously tz.ias_max_duration
+    const toZigbee: Tz.Converter[] = [
+        {
+            key: ["max_duration"],
+            convertSet: async (entity, key, value, meta) => {
+                await entity.write("ssIasWd", {maxDuration: value as number});
+                return {state: {max_duration: value}};
+            },
+            convertGet: async (entity, key, meta) => {
+                await entity.read("ssIasWd", ["maxDuration"]);
+            },
+        },
+    ];
+
+    // previously fz.ias_wd
+    const fromZigbee = [
+        {
+            cluster: "ssIasWd",
+            type: ["attributeReport", "readResponse"],
+            convert: (model, msg, publish, options, meta) => {
+                const result: KeyValueAny = {};
+                if (msg.data.maxDuration !== undefined) result.max_duration = msg.data.maxDuration;
+                return result;
+            },
+        } satisfies Fz.Converter<"ssIasWd", undefined, ["attributeReport", "readResponse"]>,
+    ];
+
+    return {toZigbee, fromZigbee, exposes, isModernExtend: true};
+}
+
 // #endregion
 
 // #region Smart Energy
